Make it possible to opt-out from otherwise recommended BOA Skynet auto-updates - fixes #557
Note that this is only first step to be included in the upcoming BOA-2.4.0 release, while BOA-2.4.1 should switch this to opt-in mode and disabled by default. It should be left for the system admin decision, but we also can't introduce opt-in mode without prior notice which should be published in the release notes.

+++ aegir/tools/system/clear.sh
@@ -20,12 +20,17 @@ fi
 if [ -e "/var/run/boa_run.pid" ] || [ -e "/var/run/daily-fix.pid" ] ; then
   sleep 1
 else
-  rm -f /tmp/*error*
-  rm -f /var/backups/BOA.sh.txt.hourly*
-  curl -L --max-redirs 10 -k -s --retry 10 --retry-delay 5 -A iCab "http://files.aegir.cc/BOA.sh.txt" -o /var/backups/BOA.sh.txt.hourly
-  bash /var/backups/BOA.sh.txt.hourly &> /dev/null
-  rm -f /var/backups/BOA.sh.txt.hourly*
-  /opt/local/bin/autoupboa
+  if [ -e "/root/.barracuda.cnf" ] ; then
+    source /root/.barracuda.cnf
+  fi
+  if [ -z "$_SKYNET_MODE" ] || [ "$_SKYNET_MODE" = "ON" ] ; then
+    rm -f /tmp/*error*
+    rm -f /var/backups/BOA.sh.txt.hourly*
+    curl -L --max-redirs 10 -k -s --retry 10 --retry-delay 5 -A iCab "http://files.aegir.cc/BOA.sh.txt" -o /var/backups/BOA.sh.txt.hourly
+    bash /var/backups/BOA.sh.txt.hourly &> /dev/null
+    rm -f /var/backups/BOA.sh.txt.hourly*
+  fi
+  bash /opt/local/bin/autoupboa
 fi
 if [ -e "/etc/resolvconf/run/interface/lo.pdnsd" ] ; then
   rm -f /etc/resolvconf/run/interface/eth*
